Add Functionality to update the PTB Repository
