+++ ptb_changelog_helper/version.py
@@ -1,0 +1,29 @@
+"""This module contains a class that represents a version number."""
+from typing import Literal, NamedTuple
+
+
+class Version(NamedTuple):
+    """This class is basically copied from PTB. We don't import it, because it's private and
+    we want to rely only on the public API."""
+
+    major: int
+    minor: int
+    micro: int
+    releaselevel: Literal["alpha", "beta", "candidate", "final"]
+    serial: int
+
+    def _rl_shorthand(self) -> str:
+        return {
+            "alpha": "a",
+            "beta": "b",
+            "candidate": "rc",
+        }[self.releaselevel]
+
+    def __str__(self) -> str:
+        version = f"{self.major}.{self.minor}"
+        if self.micro != 0:
+            version = f"{version}.{self.micro}"
+        if self.releaselevel != "final":
+            version = f"{version}{self._rl_shorthand()}{self.serial}"
+
+        return version
